feat: abstract direct model queries, create signature service
