+++ packages/auth/src/index.ts
@@ -13,6 +13,7 @@ import { createAccessTokenService } from './accessToken/service'
 import { createAccessTokenRoutes } from './accessToken/routes'
 import { createGrantRoutes } from './grant/routes'
 import { createOpenAPI } from 'openapi'
+import { createSignatureService } from './signature/service'
 
 export { JWKWithRequired } from './client/service'
 const container = initIocContainer(Config)
@@ -126,6 +127,19 @@ export function initIocContainer(
         logger: await deps.use('logger'),
         accessTokenService: await deps.use('accessTokenService'),
         clientService: await deps.use('clientService')
+      })
+    }
+  )
+
+  container.singleton(
+    'signatureService',
+    async (deps: IocContract<AppServices>) => {
+      return createSignatureService({
+        config: await deps.use('config'),
+        logger: await deps.use('logger'),
+        clientService: await deps.use('clientService'),
+        grantService: await deps.use('grantService'),
+        accessTokenService: await deps.use('accessTokenService')
       })
     }
   )
